# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Río Uruguay en Colón
  version: 0.1.0
  description: |
    API pública de solo lectura con el estado del río Uruguay en Colón (Entre Ríos).
    Fuente de verdad del contrato entre el backend FastAPI y la web.
servers:
  - url: /api
    description: Mismo origen que la web (proxy del dev server de Vite y de nginx)
  - url: http://localhost:8000
    description: API corriendo local, puerto según .env.local (API_PORT)
paths:
  /health:
    get:
      operationId: getHealth
      summary: Estado del servicio y de la conexión a la base de datos
      tags: [health]
      responses:
        "200":
          description: El proceso de la API responde; `db` indica si Postgres contesta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                ok:
                  value: { status: ok, db: ok }
                dbError:
                  value: { status: ok, db: error }
  /alturas/ultima:
    get:
      operationId: getAlturaUltima
      summary: Última altura real conocida del puerto, con tendencia y estado
      tags: [alturas]
      responses:
        "200":
          description: Última lectura disponible (INA o Prefectura).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UltimaAltura"
              examples:
                alerta:
                  value:
                    fecha_hora: "2026-09-21T12:00:00Z"
                    altura_m: 7.15
                    fuente: ina
                    tendencia_24h_m: 0.12
                    estado: alerta
        "404":
          description: Todavía no hay lecturas cargadas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                sinDatos:
                  value: { detail: "Sin datos de altura" }
  /alturas:
    get:
      operationId: listAlturasDiarias
      summary: Serie diaria de altura real del puerto (promedio por día)
      tags: [alturas]
      parameters:
        - name: desde
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: hasta
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Promedio diario de altura, ordenado por fecha ascendente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlturaDiaria"
              examples:
                rango:
                  value:
                    - { fecha: "2024-05-13", altura_m: 8.40 }
                    - { fecha: "2024-05-14", altura_m: 9.06 }
        "422":
          description: Rango inválido (desde > hasta, más de 3 años, o parámetros faltantes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                rangoMaximo:
                  value: { detail: "El rango máximo es de 3 años" }
  /pronostico:
    get:
      operationId: getPronostico
      summary: Pronóstico de Google para Colón (rango de altura y nivel de aviso)
      tags: [pronostico]
      responses:
        "200":
          description: Última emisión guardada del gauge de Colón, con su nivel de aviso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pronostico"
              examples:
                alertaProbable:
                  value:
                    emitido: "2026-09-19T20:45:00.987654Z"
                    gauge_id: hybas_6121320620
                    dias:
                      - fecha: "2026-09-19"
                        lead_dias: 0
                        caudal_m3s: 6180.0
                        altura_est_m: 5.9
                        altura_min_m: 4.9
                        altura_max_m: 6.9
                        altura_anclada_m: 6.52
                        altura_anclada_min_m: 5.52
                        altura_anclada_max_m: 7.52
                        extrapolado: false
                      - fecha: "2026-09-20"
                        lead_dias: 1
                        caudal_m3s: 12836.0
                        altura_est_m: 7.07
                        altura_min_m: 6.07
                        altura_max_m: 8.07
                        altura_anclada_m: 7.60
                        altura_anclada_min_m: 6.60
                        altura_anclada_max_m: 8.60
                        extrapolado: false
                    aviso:
                      nivel: alerta_probable
                      umbral_m3s: 11000
                      primer_dia: "2026-09-20"
                      caudal_max_m3s: 12836.0
                    anclaje:
                      aplicado: true
                      sesgo_m: 0.62
                      altura_real_m: 6.52
                      fecha_referencia: "2026-09-19"
                      motivo: null
        "503":
          description: Todavía no hay ningún pronóstico de Google guardado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                sinDatos:
                  value: { detail: "Todavía no hay un pronóstico de Google cargado" }
  /pronostico/historico:
    get:
      operationId: getPronosticoHistorico
      summary: Serie histórica del caudal pronosticado a un lead fijo, para comparar con la altura real
      tags: [pronostico]
      parameters:
        - name: lead
          in: query
          required: true
          description: Días de anticipación del pronóstico a comparar (0 a 7).
          schema:
            type: integer
            minimum: 0
            maximum: 7
        - name: desde
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: >
            Un valor por día desde `desde`, tomando (si hubo más de una
            emisión) la más reciente para ese día y ese `lead`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoricoDia"
              examples:
                serie:
                  value:
                    - fecha: "2026-09-19"
                      caudal_m3s: 6180.0
                      altura_est_m: 5.90
                      extrapolado: false
                    - fecha: "2026-09-20"
                      caudal_m3s: 12836.0
                      altura_est_m: 7.07
                      extrapolado: false
        "422":
          description: Parámetros faltantes o fuera de rango.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pronostico/aguas-arriba:
    get:
      operationId: getPronosticoAguasArriba
      summary: Pronóstico de Google para el gauge aguas arriba (zona Concordia / Salto Grande)
      tags: [pronostico]
      responses:
        "200":
          description: Última emisión guardada del gauge aguas arriba (sin nivel de aviso).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PronosticoAguasArriba"
              examples:
                sinAviso:
                  value:
                    emitido: "2026-09-19T08:15:00.123456Z"
                    gauge_id: hybas_6120865460
                    dias:
                      - fecha: "2026-09-19"
                        lead_dias: 0
                        caudal_m3s: 5800.0
                        altura_est_m: 5.55
                        altura_min_m: 4.55
                        altura_max_m: 6.55
                        extrapolado: false
        "503":
          description: Todavía no hay ningún pronóstico de Google guardado para este gauge.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                sinDatos:
                  value: { detail: "Todavía no hay un pronóstico de Google cargado" }
  /estadisticas:
    get:
      operationId: getEstadisticas
      summary: Estadísticas agregadas sobre las alturas reales y los pronósticos guardados
      description: >
        Todo se calcula sobre lo que ya está en la base (nunca llama a APIs externas
        en el request path, ver CLAUDE.md sección 3).
      tags: [estadisticas]
      responses:
        "200":
          description: Estadísticas agregadas. Los campos se degradan a listas vacías o `null`
            cuando no hay datos suficientes para calcularlos, nunca se inventan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estadisticas"
              examples:
                conDatos:
                  value:
                    percentil_hoy: { altura_m: 4.29, percentil: 78.4, ventana_dias: 365 }
                    error_pronostico: { lead_dias: 3, muestras: 120, mae_m: 0.35 }
                    dias_en_alerta:
                      - { desde: "2024-05-13", hasta: "2024-05-16", max_m: 9.06 }
                    mismo_dia_otros_anios:
                      - { anio: 2024, altura_m: 3.10 }
                      - { anio: 2025, altura_m: 5.80 }
                    eventos:
                      - { fecha: "2024-05-14", altura_m: 9.06, etiqueta: "Máximo diario registrado (INA)" }
                      - { fecha: "2025-06-18", altura_m: 7.62, etiqueta: "Crecida de referencia" }
                      - { fecha: "2026-07-09", altura_m: 4.44, etiqueta: "Costanera inundada" }
components:
  schemas:
    HealthResponse:
      type: object
      required: [status, db]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [ok]
          description: Siempre `ok` si el proceso responde.
        db:
          type: string
          enum: [ok, error]
          description: Resultado de un `SELECT 1` contra Postgres.
    UltimaAltura:
      type: object
      required: [fecha_hora, altura_m, fuente, tendencia_24h_m, estado]
      additionalProperties: false
      properties:
        fecha_hora:
          type: string
          format: date-time
          description: Momento de la lectura, en UTC.
        altura_m:
          type: number
          description: Altura real medida en el puerto, en metros sobre el cero local.
        fuente:
          type: string
          enum: [ina, prefectura]
          description: Origen del dato (INA es primario, Prefectura es respaldo).
        tendencia_24h_m:
          type: [number, "null"]
          description: >
            Diferencia con la lectura más cercana a 24 h antes (ventana de 20 a 36 h);
            `null` si no hay ninguna lectura en esa ventana.
        estado:
          type: string
          enum: [normal, evacuacion_en_seco, alerta, evacuacion]
          description: Nivel de aviso según los umbrales locales de Colón.
    AlturaDiaria:
      type: object
      required: [fecha, altura_m]
      additionalProperties: false
      properties:
        fecha:
          type: string
          format: date
          description: Día local (America/Argentina/Buenos_Aires).
        altura_m:
          type: number
          description: Promedio diario de altura real, en metros sobre el cero local.
    DiaPronostico:
      type: object
      required:
        [
          fecha,
          lead_dias,
          caudal_m3s,
          altura_est_m,
          altura_min_m,
          altura_max_m,
          altura_anclada_m,
          altura_anclada_min_m,
          altura_anclada_max_m,
          extrapolado,
        ]
      additionalProperties: false
      properties:
        fecha:
          type: string
          format: date
        lead_dias:
          type: integer
          description: Días de anticipación de este pronóstico respecto de la emisión (`emitido`).
        caudal_m3s:
          type: number
          description: >
            Caudal pronosticado por Google, en metros cúbicos por segundo. El anclaje
            (`altura_anclada_*`) nunca lo modifica.
        altura_est_m:
          type: number
          description: >
            Altura estimada del puerto a partir del caudal (curva de calibración). Se
            muestra siempre junto con `altura_min_m`/`altura_max_m`, nunca como número exacto.
        altura_min_m:
          type: number
          description: Extremo inferior del rango de altura estimada (`altura_est_m - 1,0 m`).
        altura_max_m:
          type: number
          description: Extremo superior del rango de altura estimada (`altura_est_m + 1,0 m`).
        altura_anclada_m:
          type: number
          description: >
            `altura_est_m` trasladada hacia la altura real de hoy (spec 007, C2): corrección de
            PRESENTACIÓN sobre la curva, no una recalibración. Igual a `altura_est_m` cuando
            `Pronostico.anclaje.aplicado` es `false`, o a partir del día `HORIZONTE_ANCLAJE_DIAS`.
        altura_anclada_min_m:
          type: number
          description: Igual a `altura_anclada_m` pero para `altura_min_m`.
        altura_anclada_max_m:
          type: number
          description: Igual a `altura_anclada_m` pero para `altura_max_m`.
        extrapolado:
          type: boolean
          description: >
            `true` si `caudal_m3s` supera el caudal máximo con el que se calibró la curva
            (`CAUDAL_MAX_CALIBRADO_M3S` = 15.000 m³/s): la altura estimada es una extrapolación.
    AvisoPronostico:
      type: object
      required: [nivel, umbral_m3s, primer_dia, caudal_max_m3s]
      additionalProperties: false
      properties:
        nivel:
          type: string
          enum: [sin_aviso, atencion, alerta_probable]
          description: >
            Nivel de aviso según el caudal pronosticado a 1-7 días para el gauge de Colón.
            No es el mismo umbral (ni la misma fuente) que `estado` en `/alturas/ultima`.
        umbral_m3s:
          type: [number, "null"]
          description: Umbral (m³/s) que determinó `nivel`; `null` si `nivel` es `sin_aviso`.
        primer_dia:
          type: [string, "null"]
          format: date
          description: >
            Primer día (por fecha) que cruza `umbral_m3s`; `null` si `nivel` es `sin_aviso`.
        caudal_max_m3s:
          type: [number, "null"]
          description: >
            Caudal máximo pronosticado entre los días considerados (lead 1 a 7); `null`
            si no hay ningún día en esa ventana.
    Anclaje:
      type: object
      required: [aplicado, sesgo_m, altura_real_m, fecha_referencia, motivo]
      additionalProperties: false
      properties:
        aplicado:
          type: boolean
          description: >
            `false` si no hay un día de pronóstico para hoy o no hay una altura real
            disponible para hoy; en ese caso `altura_anclada_*` queda igual a la altura
            estimada sin ajustar, y `motivo` explica por qué.
        sesgo_m:
          type: [number, "null"]
          description: >
            `altura_real_m - altura_est_m` del día de hoy; `null` si `aplicado` es `false`.
        altura_real_m:
          type: [number, "null"]
          description: Altura real promedio de hoy usada para calcular el sesgo; `null` si
            `aplicado` es `false`.
        fecha_referencia:
          type: [string, "null"]
          format: date
          description: Fecha de hoy usada como referencia; `null` si `aplicado` es `false`.
        motivo:
          type: [string, "null"]
          description: Explicación en español de por qué no se ancló; `null` si `aplicado`
            es `true`.
    Pronostico:
      type: object
      required: [emitido, gauge_id, dias, aviso, anclaje]
      additionalProperties: false
      properties:
        emitido:
          type: string
          format: date-time
          description: Momento de emisión del pronóstico usado (la última emisión guardada), en UTC.
        gauge_id:
          type: string
          description: Identificador del gauge de Google (Colón).
        dias:
          type: array
          items:
            $ref: "#/components/schemas/DiaPronostico"
          description: Solo días con `fecha >= hoy` (hora de Buenos Aires), ordenados por fecha
            ascendente.
        aviso:
          $ref: "#/components/schemas/AvisoPronostico"
        anclaje:
          $ref: "#/components/schemas/Anclaje"
    PronosticoAguasArriba:
      type: object
      required: [emitido, gauge_id, dias]
      additionalProperties: false
      properties:
        emitido:
          type: string
          format: date-time
        gauge_id:
          type: string
          description: Identificador del gauge de Google aguas arriba (zona Concordia / Salto Grande).
        dias:
          type: array
          items:
            $ref: "#/components/schemas/DiaPronostico"
    HistoricoDia:
      type: object
      required: [fecha, caudal_m3s, altura_est_m, extrapolado]
      additionalProperties: false
      properties:
        fecha:
          type: string
          format: date
        caudal_m3s:
          type: number
        altura_est_m:
          type: number
        extrapolado:
          type: boolean
    PercentilHoy:
      type: object
      required: [altura_m, percentil, ventana_dias]
      additionalProperties: false
      properties:
        altura_m:
          type: number
          description: Altura real promedio de hoy.
        percentil:
          type: number
          description: Porcentaje de días de la ventana con altura menor a la de hoy.
        ventana_dias:
          type: integer
          description: Tamaño de la ventana histórica usada (365 días).
    ErrorPronostico:
      type: object
      required: [lead_dias, muestras, mae_m]
      additionalProperties: false
      properties:
        lead_dias:
          type: integer
          description: Anticipación del pronóstico comparado (siempre 3 días).
        muestras:
          type: integer
          description: Cantidad de días con pronóstico a este lead y altura real disponibles.
        mae_m:
          type: [number, "null"]
          description: Error absoluto promedio, en metros; `null` si `muestras` es 0.
    RangoAlerta:
      type: object
      required: [desde, hasta, max_m]
      additionalProperties: false
      properties:
        desde:
          type: string
          format: date
        hasta:
          type: string
          format: date
        max_m:
          type: number
          description: Altura máxima real dentro del rango.
    MismoDiaAnio:
      type: object
      required: [anio, altura_m]
      additionalProperties: false
      properties:
        anio:
          type: integer
        altura_m:
          type: number
          description: Altura real de ese mismo día y mes, en ese año.
    Evento:
      type: object
      required: [fecha, altura_m, etiqueta]
      additionalProperties: false
      properties:
        fecha:
          type: string
          format: date
        altura_m:
          type: number
        etiqueta:
          type: string
          description: Descripción legible en español del evento de referencia.
    Estadisticas:
      type: object
      required:
        [percentil_hoy, error_pronostico, dias_en_alerta, mismo_dia_otros_anios, eventos]
      additionalProperties: false
      properties:
        percentil_hoy:
          oneOf:
            - $ref: "#/components/schemas/PercentilHoy"
            - type: "null"
          description: >
            `null` si todavía no hay una altura real promediada para hoy.
        error_pronostico:
          $ref: "#/components/schemas/ErrorPronostico"
        dias_en_alerta:
          type: array
          items:
            $ref: "#/components/schemas/RangoAlerta"
        mismo_dia_otros_anios:
          type: array
          items:
            $ref: "#/components/schemas/MismoDiaAnio"
        eventos:
          type: array
          items:
            $ref: "#/components/schemas/Evento"
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          description: >
            Mensaje de error legible. En los 422 de validación de FastAPI
            (parámetros faltantes o mal formados) es una lista de errores.
          oneOf:
            - type: string
            - type: array
              items:
                type: object